Fix x64 pointer truncation in script stack push (push_val intptr_t)
push_val2 (and the push_val macro / script_pushint) took the pushed value
as int, but the stack slot script_data.u.num is intptr_t. callfunc/callsub
save the caller's script_code and var_function POINTERS on the stack via
script_pushint((intptr_t)ptr); these were truncated to 32 bits on entry to
push_val2, then read back as truncated pointers.

This crashed any NPC using callfunc/callsub (Kafras everywhere, inn/forge
NPCs, etc.): either at the next get_com after the function returned (st->script
restored truncated, e.g. 0x58886510 instead of 0x555558886510) or later in
script_free_stack -> script_free_vars (truncated var_function, e.g. 0x66fa8520).

Change push_val2's val parameter (and the push_val forward declaration) from
int to intptr_t so pointers survive a round-trip on the script stack.

+++ src/map/script.c
@@ -250,7 +250,7 @@ static struct linkdb_node *sleep_db;
  * [Jvg^Cv (Kv)
  *------------------------------------------*/
 const char* parse_subexpr(const char* p,int limit);
-void push_val(struct script_stack *stack,int type,int val);
+void push_val(struct script_stack *stack,int type,intptr_t val);
 int run_func(struct script_state *st);
 
 int mapreg_setreg(int num,int val);
@@ -2383,7 +2383,7 @@ void stack_expand(struct script_stack* stack)
 #define push_val(stack,type,val) push_val2(stack, type, val, NULL)
 
 /// Pushes a value into the stack (with reference)
-void push_val2(struct script_stack* stack, int type, int val, struct linkdb_node** ref)
+void push_val2(struct script_stack* stack, int type, intptr_t val, struct linkdb_node** ref)
 {
 	if( stack->sp >= stack->sp_max )
 		stack_expand(stack);
